+++ src/molecule/shellpair.h
@@ -48,7 +48,7 @@ class ShellPair {
   public:
     ShellPair() { }
     ShellPair(const std::array<std::shared_ptr<const Shell>, 2>& shells, const std::array<int, 2>& offset, const std::pair<int, int>& shell_ind, const double thresh = 1e-10);
-    bool is_neighbour(std::shared_ptr<const ShellPair> sp, const int ws) const;
+    bool is_neighbour(std::shared_ptr<const ShellPair> sp, const double ws) const;
 
     const std::array<std::shared_ptr<const Shell>, 2>& shells() const { return shells_; }
     std::shared_ptr<const Shell> shell(const int i) const { assert(i==0 || i==1); return shells_[i]; }
